Time: 44 ms (89.93%), Space: 18.8 MB (77.93%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -6,7 +6,7 @@ public:
         double  sum=0;
         for(int i=0;i<piles.size();i++)
         {
-           sum+=ceil(piles[i]/(double)h); 
+           sum+=ceil((double)piles[i]/(double)h); 
         }
         return sum;
     }
